refactor: split main.go into connect.go / notify.go / serve.go (#20 I1)
Pure file move — no function changed, no behavior changed. main.go was 2617
lines; declarations relocate along the concern boundaries the file already
had:

- connect.go (1216): connectOpts and the flag mutexes, cmdConnect,
  runConnect, cmdSetup, deploy-state construction, binary
  preparation/download, tunnel verification, and the whole Codex section
  (runConnectCodex + bridge helpers).
- notify.go (568): the detect-install adapter table and runner,
  connectNotifySetup, notify deploy-state merging, Claude hooks
  configuration, nonce sync/registration, health probe, and cmdNotify with
  its Codex payload parser.
- serve.go (86): cmdServe.
- main.go (788): entrypoint, usage, flag helpers, paste/install/uninstall,
  doctor, status, service, x11-bridge dispatch.

Every block moved with its doc comments and section banners; import blocks
were re-derived per file. Verified move-only: top-level declaration count is
identical before and after (67), and the full race suite, vet, staticcheck,
and GOOS=windows/linux builds all pass untouched.

This lands the file-move reading of I1 — and main.go ends at 788 lines,
under the issue's original ~800 target after all. CLAUDE.md's three symbol
pointers (cmdConnect, runConnectCodex, deploy steps N3/N4) now name the new
files.
